amiga.c - moved GlobalStuffChar declaration from lsprintf.c to amiga.c           moved LSprintf() function from lsprintf.c to amiga.c amiga.h - added KPrintF definition for OS4. libinit.c - Added copyright string (defined in version/dopus_version.h) module_deps.h - Changed version string definitions to use definitions in                 version/dopus_version.h XADopus.c - Fixed format size in sprintf function call (%d to %ld). makefiles - added date defines for use in program version strings. added newlines at end of files to eliminate compile warnings. removed unused files lsprinf.c & version.h

+++ trunk/source/Modules/xadopus/XADopus.c
@@ -1036,7 +1036,7 @@ int LIBFUNC SAVEDS ASM L_Module_Entry(
 					AddPart(arcname,Entry->name,512);
 				}
 			}
-			sprintf(buf,"command source %d ScanDir %s",data.destp->lister,data.destp->path);
+			sprintf(buf,"command source %ld ScanDir %s",(ULONG)data.destp->lister,data.destp->path);
 			DC_CALL4(infoptr, dc_SendCommand,
 				DC_REGA0, IPCDATA(ipc),
 				DC_REGA1, buf,
